Fix databinding, merge security test into main test class
